Fix for perf tracing integer overflow regression
The `interval` variable is of type `uint64_t`, so the expression `interval * interval` also evaluates to a value of type `uint64_t`, meaning the result of the multiplication will be truncated if `interval` is `1llu << 32` or greater, regardless of the fact that the result is stored in a variable of type `__uint128_t`. Casting one side of the multiplication to the larger type fixes the problem.

+++ ProjFS.Mac/PrjFSKext/PrjFSKext/PerformanceTracing.cpp
@@ -46,7 +46,7 @@ void PerfTracing_RecordSample(PrjFSPerfCounter counter, uint64_t startTime, uint
     {
         atomic_fetch_add(&probe->sum, interval);
 
-        __uint128_t intervalSquared = interval * interval;
+        __uint128_t intervalSquared = static_cast<__uint128_t>(interval) * interval;
         atomic_fetch_add(&probe->sumSquares, intervalSquared);
         
         // Update minimum sample if necessary
